requirements/main.txt
Pillow
torch
torchvision
matplotlib
opencv-python
albumentations
transformers
hydra-core
